Module Functoria.Engine

Functoria engine.

val if_keys : Impl.abstract -> Key.Set.t

if_keys t is the set of if keys in the graph t.

val all_keys : Impl.abstract -> Key.Set.t

all_keys t is the set of keys in the graph t.

val packages : Impl.abstract -> Package.t list Key.value

packages t is the set of packages in the graph t.

val install : Info.t -> Impl.abstract -> Install.t Key.value

install i t is the set of files installed by the graph t.

val files : Info.t -> Impl.abstract -> Fpath.Set.t

files i t is the list of files generated configure-time.

val dune : Info.t -> Impl.abstract -> Dune.stanza list

dune i t is the list of dune stanzas needed to build the project t with the build information i.

Triggering Hooks

type t = Device.Graph.t

The type for key graphs.

val configure : Info.t -> t -> unit Action.t

configure i t calls all the configuration hooks for each of the implementations appearing in t, in topological order. Use the build information i.

val connect : ?init:'a Impl.t list -> Info.t -> t -> unit Action.t

connect ?init i t generates the connect functions in main.ml, for each of the implementations appearing t, in topological order. Use build information i.