Module Tool.Make

Parameters

module P : S

Signature

val run : unit -> unit

Run the configuration builder. This should be called exactly once to run the configuration builder: command-line arguments will be parsed, and some code will be generated and compiled.

val run_with_argv : ?help_ppf:Stdlib.Format.formatter -> ?err_ppf:Stdlib.Format.formatter -> string array -> unit

run_with_argv a is the same as run but parses a instead of the process command line arguments. It also allows to set the error and help channels using help_ppf and err_ppf.