Module type Functoria_app.S

S is the signature that application builders have to provide.

val prelude : string

Prelude printed at the beginning of main.ml.

It should put in scope:

a run function of type 'a t -> 'a

a return function of type 'a -> 'a t

a >>= operator of type 'a t -> ('a -> 'b t) -> 'b t

val name : string

Name of the custom DSL.

val packages : Functoria.package list

The packages to load when compiling the configuration file.

val ignore_dirs : string list

Directories to ignore when compiling the configuration file.

val version : string

Version of the custom DSL.

val create : Functoria.job Functoria.impl list -> Functoria.job Functoria.impl

create jobs is the top-level job in the custom DSL which will execute the given list of job.